+++ lib/src/requests/delete_custom_xml_part_online_request.dart
@@ -109,6 +109,7 @@ class DeleteCustomXmlPartOnlineRequest implements RequestBase {
     }
 
     for (final _fileContentPart in _fileContentParts) {
+        _fileContentPart.encryptPassword(_apiClient);
         if (_fileContentPart.source == 'Request') {
             _bodyParts.add(ApiRequestPart(_fileContentPart.content!, 'application/octet-stream', name: _fileContentPart.reference));
         }