+++ lib/src/requests/get_document_drawing_object_image_data_online_request.dart
@@ -91,6 +91,7 @@ class GetDocumentDrawingObjectImageDataOnlineRequest implements RequestBase {
     }
 
     for (final _fileContentPart in _fileContentParts) {
+        _fileContentPart.encryptPassword(_apiClient);
         if (_fileContentPart.source == 'Request') {
             _bodyParts.add(ApiRequestPart(_fileContentPart.content!, 'application/octet-stream', name: _fileContentPart.reference));
         }